examples/python_requirements.txt
pyyaml
numpy
matplotlib